Extracts shared ResourceTestHelpers in MCP test project
MakeUninitializedContext and MakeReadContext were duplicated verbatim
across four resource test classes; centralised into a single internal
static helper.
