+++ src/Conjecture.Mcp.Tests/AspNetCoreEFCoreSetupResourceTests.cs
@@ -1,9 +1,6 @@
 // Copyright (c) 2026 Kim Ommundsen. Licensed under the MPL-2.0.
 // See LICENSE.txt in the project root or https://mozilla.org/MPL/2.0/
 
-using System.Reflection;
-using System.Runtime.CompilerServices;
-
 using Conjecture.Mcp.Resources;
 
 using ModelContextProtocol.Protocol;
@@ -13,30 +10,12 @@ namespace Conjecture.Mcp.Tests;
 
 public class AspNetCoreEFCoreSetupResourceTests
 {
-    private static RequestContext<T> MakeUninitializedContext<T>()
-        where T : class
-    {
-        Type contextType = typeof(RequestContext<T>);
-        object ctx = RuntimeHelpers.GetUninitializedObject(contextType);
-        return (RequestContext<T>)ctx;
-    }
-
-    private static RequestContext<ReadResourceRequestParams> MakeReadContext(string uri)
-    {
-        Type contextType = typeof(RequestContext<ReadResourceRequestParams>);
-        object ctx = RuntimeHelpers.GetUninitializedObject(contextType);
-        FieldInfo field = contextType.GetField(
-            "<Params>k__BackingField",
-            BindingFlags.NonPublic | BindingFlags.Instance)!;
-        field.SetValue(ctx, new ReadResourceRequestParams { Uri = uri });
-        return (RequestContext<ReadResourceRequestParams>)ctx;
-    }
 
     [Fact]
     public async Task AspNetCoreEFCoreSetupResource_ExistsInListResources()
     {
         ListResourcesResult result = await ApiReferenceResources.HandleListResources(
-            MakeUninitializedContext<ListResourcesRequestParams>(), default);
+            ResourceTestHelpers.MakeUninitializedContext<ListResourcesRequestParams>(), default);
 
         Assert.Contains(result.Resources, r => r.Uri == "conjecture://api/aspnetcore-efcore-setup");
     }
@@ -49,7 +28,7 @@ public class AspNetCoreEFCoreSetupResourceTests
     public async Task AspNetCoreEFCoreSetupResource_ExistsAndContainsExpectedSnippets(string expectedTerm)
     {
         RequestContext<ReadResourceRequestParams> context =
-            MakeReadContext("conjecture://api/aspnetcore-efcore-setup");
+            ResourceTestHelpers.MakeReadContext("conjecture://api/aspnetcore-efcore-setup");
 
         ReadResourceResult result = await ApiReferenceResources.HandleReadResource(
             context, default);
@@ -62,7 +41,7 @@ public class AspNetCoreEFCoreSetupResourceTests
     public async Task AspNetCoreEFCoreSetupResource_ContentMentionsTestHostWiring()
     {
         RequestContext<ReadResourceRequestParams> context =
-            MakeReadContext("conjecture://api/aspnetcore-efcore-setup");
+            ResourceTestHelpers.MakeReadContext("conjecture://api/aspnetcore-efcore-setup");
 
         ReadResourceResult result = await ApiReferenceResources.HandleReadResource(
             context, default);
